requirements.txt
absl-py==1.0.0
attrs==21.4.0
cycler==0.11.0
fonttools==4.29.0
imutils==0.5.4
kiwisolver==1.3.2
matplotlib==3.5.1
mediapipe==0.8.6
numpy==1.21.4
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.5
pathlib==1.0.1
Pillow==9.0.0
protobuf==3.19.3
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
six==1.16.0
tk==0.1.0
